# user-service/src/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CustomApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {},
					"status_code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status_code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StateLGA": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"type": "string"
					},
					"lga": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"lga"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"stateLgaId": {
						"type": "number",
						"format": "double"
					},
					"profilePicture": {
						"type": "string"
					},
					"tier": {
						"type": "number",
						"format": "double"
					},
					"pin": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					},
					"password_reset_token": {
						"type": "string"
					},
					"token_expires_at": {
						"type": "string",
						"format": "date-time"
					},
					"isValidated": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"isEnabled": {
						"type": "boolean"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"lastUpdatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"stateLga": {
						"$ref": "#/components/schemas/StateLGA"
					}
				},
				"required": [
					"email",
					"password",
					"firstName",
					"lastName",
					"address",
					"phoneNumber",
					"stateLgaId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FetchProfileResponseDTO": {
				"properties": {
					"user": {
						"allOf": [
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PasswordResetResponseDTO": {
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"isSuccess",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PasswordResetRequestDTO": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SetNewPasswordRequestDTO": {
				"properties": {
					"password_reset_token": {
						"type": "string"
					},
					"new_password": {
						"type": "string"
					}
				},
				"required": [
					"password_reset_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginUserDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailVerificationDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "df_wallet",
		"version": "1.0.0",
		"description": "This is the DF Wallet RESTful API",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "IfeoluwaOkesola ",
			"email": "99264083+IfeoluwaOkesola@users.noreply.github.com"
		}
	},
	"paths": {
		"/withdrawal-accounts": {
			"post": {
				"operationId": "CreateNewWithdrawalAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Withdrawal Bank Accounts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"get": {
				"operationId": "ListWithdrawalAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Withdrawal Bank Accounts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/withdrawal-accounts/{account_id}": {
			"delete": {
				"operationId": "DeleteWithdrawalAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Withdrawal Bank Accounts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "account_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"get": {
				"operationId": "ShowWithdrawalAccountDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Withdrawal Bank Accounts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "account_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "FetchProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FetchProfileResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"User Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User Profile"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/password/reset": {
			"post": {
				"operationId": "PerformPasswordResetRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PasswordResetResponseDTO"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "boniface@developersfoundry.com"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Password Reset"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PasswordResetRequestDTO"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "SetNewPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PasswordResetResponseDTO"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "boniface@developersfoundry.com"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Password Reset"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetNewPasswordRequestDTO"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserDTO"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUserDTO"
							}
						}
					}
				}
			}
		},
		"/auth/email/verification": {
			"put": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": true
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmailVerificationDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:2025/"
		}
	]
}